more 'pythonic' module structure move code to directory with more generic module name (plan to factor out general engine code later)
