+++ cardshark/engine.py
@@ -1,8 +1,8 @@
 # other Coup AI stuff
-import Logging
-from Cards import *
-from Player import *
-from named_object import NamedObject
+from cardshark import cards
+from cardshark.logging import *
+from cardshark.player import Player
+from cardshark.named_object import NamedObject
 
 # TF agents stuff
 from tf_agents.specs import BoundedArraySpec
@@ -14,6 +14,7 @@ import itertools
 import typing
 import numpy as np
 from abc import ABC, abstractmethod
+from enum import Enum
 
 ## number of cards dealt to each player
 MAX_CARDS = 2
@@ -40,7 +41,7 @@ for i, name in enumerate(actionNames):
         actionString = actionString + " " + name 
     else: actionString = name
 ## used for converting action name to id
-actionEnum   = Enum("actionEnum", actionString)
+actionEnum = Enum("actionEnum", actionString)
 DEBUG("actionString:", actionString)
 
 
@@ -86,7 +87,7 @@ class Game(py_environment.PyEnvironment, NamedObject):
         observationSpecMin_NP = np.ndarray((nPlayers, 1 + MAX_CARDS), dtype=np.float32)
         observationSpecMin_NP[:] = 0.0
         observationSpecMax_NP[:, 0] = 12 # <- 12 is the max number of coins a player can have
-        observationSpecMax_NP[:, 1:] = len(cards.keys()) +1 # <- one index for each possible card + one for face down card 
+        observationSpecMax_NP[:, 1:] = len(cards.cards.keys()) +1 # <- one index for each possible card + one for face down card 
         
         self.DEBUG("observationSpecMin_NP: ", observationSpecMin_NP)
         self.DEBUG("observationSpecMax_NP: ", observationSpecMax_NP)
@@ -108,7 +109,7 @@ class Game(py_environment.PyEnvironment, NamedObject):
 
         ## initialise the deck
         self.TRACE("  Creating deck")
-        self.Deck = Deck()
+        self.Deck = cards.Deck()
 
         self._maxSteps = maxSteps
         self._reset()
@@ -477,7 +478,7 @@ class Game(py_environment.PyEnvironment, NamedObject):
         observation[0,0] = self.playerList[playerIdx].Coins
         ## can always see own cards
         for i in range(MAX_CARDS):
-            observation[0, 1 + i] = cardEnum[self.playerList[playerIdx].Cards[i]].value
+            observation[0, 1 + i] = cards.cardEnum[self.playerList[playerIdx].Cards[i]].value
 
         ## for the rest of the observation we fill up the equivalent for other players
         for otherPlayerCounter in range(1, self.nPlayers):
@@ -488,7 +489,7 @@ class Game(py_environment.PyEnvironment, NamedObject):
             ## can only see other players cards if they are dead                
             for i in range(MAX_CARDS):
                 if self.playerList[otherPlayerIdx].CardStates[i] == "Dead":
-                    observation[otherPlayerCounter, 1 + i] = cardEnum[self.playerList[otherPlayerIdx].Cards[i]].value     
+                    observation[otherPlayerCounter, 1 + i] = cards.cardEnum[self.playerList[otherPlayerIdx].Cards[i]].value     
                 else:   
                     observation[otherPlayerCounter, 1 + i] = 0.0
